Fixed wait_ms()
This function was setup for a different internal OSSC speed. The speed
the PIC is operating at now is 1.5MHz, so this function is not adjusted
to operate properly at this speed. Zack and I also introduced a
wait_sec(), which now allows us to wait in seconds without mental
strain.